MoviePricingCategory class created

+++ src/com/twu/refactor/MoviePricingCategory.java
@@ -1,8 +1,5 @@
 package com.twu.refactor;
 
-/**
- * Created by suparnad on 3/17/2015.
- */
 public enum MoviePricingCategory {
     KIDS(1.5, 1.5, 3),
     REGULAR(2, 1.5, 2),
@@ -10,15 +7,15 @@ public enum MoviePricingCategory {
 
     private double upFrontCost;
     private double perDayCost;
-    private int upFrontValidity;
+    private double upFrontValidity;
 
-    MoviePricingCategory(double upFrontCost, double perDayCost, int upFrontValidity) {
+    MoviePricingCategory(double upFrontCost, double perDayCost, double upFrontValidity){
         this.upFrontCost = upFrontCost;
         this.perDayCost = perDayCost;
         this.upFrontValidity = upFrontValidity;
     }
 
-    public double getCostFor(int numberOfDays) {
+    public double getCostFor(double numberOfDays) {
         double rentalAmount = upFrontCost;
         if (numberOfDays > upFrontValidity)
             rentalAmount += (numberOfDays - upFrontValidity) * perDayCost;
